+++ ActiveQuery.php
@@ -295,7 +295,18 @@ class ActiveQuery extends Query implements ActiveQueryInterface
                 ],
             ];
         }
-        $result = $this->createCommand($db)->search($options);
+
+        $command = $this->createCommand($db);
+        $result = $command->search($options);
+        if ($result === false) {
+            throw new Exception('Elasticsearch search query failed.', [
+                'index' => $command->index,
+                'type' => $command->type,
+                'query' => $command->queryParts,
+                'options' => $command->options,
+            ]);
+        }
+
         // TODO implement with() for asArray
         if (!empty($result['hits']['hits']) && !$this->asArray) {
             $models = $this->createModels($result['hits']['hits']);
@@ -316,11 +327,14 @@ class ActiveQuery extends Query implements ActiveQueryInterface
      */
     public function column($field, $db = null)
     {
-        if ($field == '_id') {
+        if ($field === '_id') {
             $command = $this->createCommand($db);
             $command->queryParts['fields'] = [];
             $command->queryParts['_source'] = false;
             $result = $command->search();
+            if ($result === false) {
+                throw new Exception('Elasticsearch search query failed.');
+            }
             if (empty($result['hits']['hits'])) {
                 return [];
             }