+++ QueryBuilder.php
@@ -47,25 +47,13 @@ class QueryBuilder extends BaseObject
     {
         $parts = [];
 
-        if ($query->fields === []) {
-            $parts['fields'] = [];
-        } elseif ($query->fields !== null) {
-            $fields = [];
-            $scriptFields = [];
-            foreach ($query->fields as $key => $field) {
-                if (is_int($key)) {
-                    $fields[] = $field;
-                } else {
-                    $scriptFields[$key] = $field;
-                }
-            }
-            if (!empty($fields)) {
-                $parts['fields'] = $fields;
-            }
-            if (!empty($scriptFields)) {
-                $parts['script_fields'] = $scriptFields;
-            }
+        if ($query->storedFields !== null) {
+            $parts['stored_fields'] = $query->storedFields;
         }
+        if ($query->scriptFields !== null) {
+            $parts['script_fields'] = $query->scriptFields;
+        }
+
         if ($query->source !== null) {
             $parts['_source'] = $query->source;
         }
@@ -82,27 +70,11 @@ class QueryBuilder extends BaseObject
             $parts['explain'] = $query->explain;
         }
 
-        if (empty($query->query)) {
-            $parts['query'] = ["match_all" => (object)[]];
-        } else {
+        $whereQuery = $this->buildQueryFromWhere($query->where);
+        if ($whereQuery) {
+            $parts['query'] = $whereQuery;
+        } else if ($query->query) {
             $parts['query'] = $query->query;
-        }
-
-        $whereFilter = $this->buildCondition($query->where);
-        if (is_string($query->filter)) {
-            if (empty($whereFilter)) {
-                $parts['filter'] = $query->filter;
-            } else {
-                $parts['filter'] = '{"and": [' . $query->filter . ', ' . Json::encode($whereFilter) . ']}';
-            }
-        } elseif ($query->filter !== null) {
-            if (empty($whereFilter)) {
-                $parts['filter'] = $query->filter;
-            } else {
-                $parts['filter'] = ['and' => [$query->filter, $whereFilter]];
-            }
-        } elseif (!empty($whereFilter)) {
-            $parts['filter'] = $whereFilter;
         }
 
         if (!empty($query->highlight)) {
@@ -119,6 +91,9 @@ class QueryBuilder extends BaseObject
         }
         if (!empty($query->postFilter)) {
             $parts['post_filter'] = $query->postFilter;
+        }
+        if (!empty($query->collapse)) {
+            $parts['collapse'] = $query->collapse;
         }
 
         $sort = $this->buildOrderBy($query->orderBy);
@@ -170,6 +145,20 @@ class QueryBuilder extends BaseObject
         return $orders;
     }
 
+    public function buildQueryFromWhere($condition) {
+        $where = $this->buildCondition($condition);
+        if ($where) {
+            $query = [
+                'constant_score' => [
+                    'filter' => $where,
+                ],
+            ];
+            return $query;
+        } else {
+            return null;
+        }
+    }
+
     /**
      * Parses the condition specification and generates the corresponding SQL expression.
      *
@@ -182,8 +171,8 @@ class QueryBuilder extends BaseObject
     {
         static $builders = [
             'not' => 'buildNotCondition',
-            'and' => 'buildAndCondition',
-            'or' => 'buildAndCondition',
+            'and' => 'buildBoolCondition',
+            'or' => 'buildBoolCondition',
             'between' => 'buildBetweenCondition',
             'not between' => 'buildBetweenCondition',
             'in' => 'buildInCondition',
@@ -226,7 +215,7 @@ class QueryBuilder extends BaseObject
 
     private function buildHashCondition($condition)
     {
-        $parts = [];
+        $parts = $emptyFields = [];
         foreach ($condition as $attribute => $value) {
             if ($attribute == '_id') {
                 if ($value === null) { // there is no null pk
@@ -236,10 +225,10 @@ class QueryBuilder extends BaseObject
                 }
             } else {
                 if (is_array($value)) { // IN condition
-                    $parts[] = ['in' => [$attribute => $value]];
+                    $parts[] = ['terms' => [$attribute => $value]];
                 } else {
                     if ($value === null) {
-                        $parts[] = ['missing' => ['field' => $attribute, 'existence' => true, 'null_value' => true]];
+                        $emptyFields[] = [ 'exists' => [ 'field' => $attribute ] ];
                     } else {
                         $parts[] = ['term' => [$attribute => $value]];
                     }
@@ -247,7 +236,11 @@ class QueryBuilder extends BaseObject
             }
         }
 
-        return count($parts) === 1 ? $parts[0] : ['and' => $parts];
+        $query = [ 'must' => $parts ];
+        if ($emptyFields) {
+            $query['must_not'] = $emptyFields;
+        }
+        return [ 'bool' => $query ];
     }
 
     private function buildNotCondition($operator, $operands)
@@ -261,12 +254,24 @@ class QueryBuilder extends BaseObject
             $operand = $this->buildCondition($operand);
         }
 
-        return [$operator => $operand];
+        return [
+            'bool' => [
+                'must_not' => $operand,
+            ],
+        ];
     }
 
-    private function buildAndCondition($operator, $operands)
+    private function buildBoolCondition($operator, $operands)
     {
         $parts = [];
+        if ($operator === 'and') {
+            $clause = 'must';
+        } else if ($operator === 'or') {
+            $clause = 'should';
+        } else {
+            throw new InvalidParamException("Operator should be 'or' or 'and'");
+        }
+
         foreach ($operands as $operand) {
             if (is_array($operand)) {
                 $operand = $this->buildCondition($operand);
@@ -275,10 +280,14 @@ class QueryBuilder extends BaseObject
                 $parts[] = $operand;
             }
         }
-        if (!empty($parts)) {
-            return [$operator => $parts];
+        if ($parts) {
+            return [
+                'bool' => [
+                    $clause => $parts,
+                ]
+            ];
         } else {
-            return [];
+            return null;
         }
     }
 
@@ -289,12 +298,12 @@ class QueryBuilder extends BaseObject
         }
 
         list($column, $value1, $value2) = $operands;
-        if ($column == '_id') {
+        if ($column === '_id') {
             throw new NotSupportedException('Between condition is not supported for the _id field.');
         }
         $filter = ['range' => [$column => ['gte' => $value1, 'lte' => $value2]]];
-        if ($operator == 'not between') {
-            $filter = ['not' => $filter];
+        if ($operator === 'not between') {
+            $filter = ['bool' => ['must_not'=>$filter]];
         }
 
         return $filter;
@@ -302,8 +311,8 @@ class QueryBuilder extends BaseObject
 
     private function buildInCondition($operator, $operands)
     {
-        if (!isset($operands[0], $operands[1])) {
-            throw new InvalidParamException("Operator '$operator' requires two operands.");
+        if (!isset($operands[0], $operands[1]) || !is_array($operands)) {
+            throw new InvalidParamException("Operator '$operator' requires array of two operands: column and values");
         }
 
         list($column, $values) = $operands;
@@ -330,37 +339,52 @@ class QueryBuilder extends BaseObject
                 unset($values[$i]);
             }
         }
-        if ($column == '_id') {
+        if ($column === '_id') {
             if (empty($values) && $canBeNull) { // there is no null pk
                 $filter = ['terms' => ['_uid' => []]]; // this condition is equal to WHERE false
             } else {
                 $filter = ['ids' => ['values' => array_values($values)]];
                 if ($canBeNull) {
                     $filter = [
-                        'or' => [
-                            $filter,
-                            ['missing' => ['field' => $column, 'existence' => true, 'null_value' => true]]
-                        ]
+                        'bool' => [
+                            'should' => [
+                                $filter,
+                                'bool' => ['must_not' => ['exists' => ['field'=>$column]]],
+                            ],
+                        ],
                     ];
                 }
             }
         } else {
             if (empty($values) && $canBeNull) {
-                $filter = ['missing' => ['field' => $column, 'existence' => true, 'null_value' => true]];
+                $filter = [
+                    'bool' => [
+                        'must_not' => [
+                            'exists' => [ 'field' => $column ],
+                        ]
+                    ]
+                ];
             } else {
-                $filter = ['in' => [$column => array_values($values)]];
+                $filter = [ 'terms' => [$column => array_values($values)] ];
                 if ($canBeNull) {
                     $filter = [
-                        'or' => [
-                            $filter,
-                            ['missing' => ['field' => $column, 'existence' => true, 'null_value' => true]]
-                        ]
+                        'bool' => [
+                            'should' => [
+                                $filter,
+                                'bool' => ['must_not' => ['exists' => ['field'=>$column]]],
+                            ],
+                        ],
                     ];
                 }
             }
         }
-        if ($operator == 'not in') {
-            $filter = ['not' => $filter];
+
+        if ($operator === 'not in') {
+            $filter = [
+                'bool' => [
+                    'must_not' => $filter,
+                ],
+            ];
         }
 
         return $filter;
@@ -380,7 +404,7 @@ class QueryBuilder extends BaseObject
         }
 
         list($column, $value) = $operands;
-        if ($column == '_id') {
+        if ($column === '_id') {
             $column = '_uid';
         }