+++ tests/CommandTest.php
@@ -491,4 +491,23 @@ class CommandTest extends TestCase
 
         $this->assertTrue($actualResult);
     }
+
+    /*
+    public function testIndexStats()
+    {
+        $cmd = $this->connection->createCommand();
+        if (!$cmd->indexExists('yii2test2')) {
+            $cmd->createIndex('yii2test2');
+        }
+        $stats = $cmd->getIndexStats();
+        $this->assertArrayHasKey('_all', $stats, print_r(array_keys($stats), true));
+        $this->assertArrayHasKey('indices', $stats, print_r(array_keys($stats), true));
+        $this->assertArrayHasKey('yii2test2', $stats['indices'], print_r(array_keys($stats['indices']), true));
+
+        $stats = $cmd->getIndexStats('yii2test2');
+        $this->assertArrayHasKey('_all', $stats, print_r(array_keys($stats), true));
+        $this->assertArrayHasKey('indices', $stats, print_r(array_keys($stats), true));
+        $this->assertArrayHasKey('yii2test2', $stats['indices'], print_r(array_keys($stats['indices']), true));
+    }
+    */
 }